2 new scenes with particle motion capture ppl
MariaCamo & CeciStarborn were created using quite a few particles. If this is too much load for the whole program, perhaps we can drop into a hole and see them. Although I do think it would be amazing to have Ceci in the sky as a cloud dancer. I love the new shader on the terrain with the rainbow effect. I was getting an error in regards to the tree shader but didn't address it.